added column to show distribution data and moved ReputationItemDelegate to GxsIdDetails
