+++ retroshare-gui/src/gui/gxsforums/GxsForumThreadWidget.ui
@@ -6,7 +6,7 @@
    <rect>
     <x>0</x>
     <y>0</y>
-    <width>851</width>
+    <width>1217</width>
     <height>721</height>
    </rect>
   </property>
@@ -236,6 +236,11 @@
          <column>
           <property name="text">
            <string>Date</string>
+          </property>
+         </column>
+         <column>
+          <property name="text">
+           <string>Distribution</string>
           </property>
          </column>
          <column>
